+++ src/pi_auto_api/externals/__init__.py
@@ -1,0 +1,3 @@
+"""Package for external service client utilities (Docassemble, DocuSign, etc.)."""
+
+# Empty file to make this a package